Remove incorrect paragraphs regarding float/double encoding.
Roblox always sets a number attribute as a Double. However, when adding
a new number attribute from Studio's properties panel, it will be
initialized as a Float.

Additionally, the Float type will be decoded successfully, and the type
will be retained as long as the attribute isn't set. This behavior also
applies to several other types, but because Roblox's implementation
does not expose them, they are not included in the spec. On the other
hand, because the Float type *is* exposed, it is retained in the spec.